Setup tests working from local.:
